Extract transition duration function to mixin
