+++ addon/components/as-side-panel.js
@@ -1,23 +1,12 @@
 import Ember from 'ember';
 import KeyEventsMixin from 'ember-cli-paint/mixins/key-events';
+import TransitionDurationMixin from 'ember-cli-paint/mixins/transition-duration';
 import InboundActions from 'ember-component-inbound-actions/inbound-actions';
 
-export default Ember.Component.extend(KeyEventsMixin, InboundActions, {
+export default Ember.Component.extend(KeyEventsMixin, TransitionDurationMixin, InboundActions, {
   classNameBindings: [':side-panel', 'isActive:active', 'isDrawerActive:drawer-active'],
   tagName: 'section',
   isDrawerActive: false,
-
-  transitionDuration: Ember.computed(function() {
-    var cssDuration = this.$().css('transition-duration');
-    var cssDelay = this.$().css('transition-delay');
-    var duration = parseFloat(cssDuration) + parseFloat(cssDelay);
-
-    if (cssDuration.indexOf('ms') !== -1) {
-      return duration;
-    } else {
-      return duration * 1000;
-    }
-  }).volatile(),
 
   onDidInsertElement: function() {
     Ember.run.schedule('afterRender', () => {
