+++ app/mixins/transition-duration.js
@@ -1,0 +1,4 @@
+import Ember from 'ember';
+import TransitionDurationMixin from 'ember-cli-paint/mixins/transition-duration';
+
+export default TransitionDurationMixin;
